backend/requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=openblas
brotli=1.0.9=h6c40b1e_8
brotli-bin=1.0.9=h6c40b1e_8
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.7.2=hecd8cb5_0
contourpy=1.2.0=py312ha357a0b_0
cycler=0.11.0=pyhd3eb1b0_0
expat=2.6.3=h6d0c2b6_0
ezkl=14.0.0=pypi_0
filelock=3.13.1=py312hecd8cb5_0
fonttools=4.51.0=py312h6c40b1e_0
freetype=2.12.1=hd8bbffd_0
fsspec=2024.6.1=py312hecd8cb5_0
jinja2=3.1.4=py312hecd8cb5_0
jpeg=9e=h46256e1_3
kiwisolver=1.4.4=py312hcec6c5f_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libabseil=20240116.2=cxx17_h548131f_0
libbrotlicommon=1.0.9=h6c40b1e_8
libbrotlidec=1.0.9=h6c40b1e_8
libbrotlienc=1.0.9=h6c40b1e_8
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_1
libffi=3.4.4=hecd8cb5_1
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libopenblas=0.3.21=h54e7dc3_0
libpng=1.6.39=h6c40b1e_0
libprotobuf=4.25.3=h34eed0b_0
libtiff=4.5.1=hcec6c5f_0
libwebp-base=1.3.2=h6c40b1e_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_1
markupsafe=2.1.3=py312h6c40b1e_0
matplotlib=3.9.2=py312hecd8cb5_0
matplotlib-base=3.9.2=py312ha7ebc0d_0
mpmath=1.3.0=py312hecd8cb5_0
ncurses=6.4=hcec6c5f_0
networkx=3.3=py312hecd8cb5_0
nomkl=3.0=0
numpy=1.26.4=py312h7d6adbd_0
numpy-base=1.26.4=py312hbb3573c_0
openjpeg=2.5.2=hbf2204d_0
openssl=3.0.15=h46256e1_0
packaging=24.1=py312hecd8cb5_0
pillow=10.4.0=py312h46256e1_0
pip=24.2=py312hecd8cb5_0
pyparsing=3.1.2=py312hecd8cb5_0
python=3.12.5=hcd54a6c_1
python-dateutil=2.9.0post0=py312hecd8cb5_2
pytorch=2.3.0=cpu_py312h781a981_1
readline=8.2=hca72f7f_0
setuptools=75.1.0=py312hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h6c40b1e_0
sympy=1.13.2=py312hecd8cb5_0
tk=8.6.14=h4d00af3_0
torchvision=0.18.1=cpu_py312h937ada3_0
tornado=6.4.1=py312h46256e1_0
typing_extensions=4.11.0=py312hecd8cb5_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py312h6c40b1e_0
wheel=0.44.0=py312hecd8cb5_0
xz=5.4.6=h6c40b1e_1
zlib=1.2.13=h4b97444_1
zstd=1.5.5=hc035e20_2
